webinar-materials/starter-kits/resume-analyzer/requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
pydantic==2.11.7
python-dotenv==1.1.1
pandas==2.3.2
